more changes

+++ ft_strncmp.c
@@ -3,21 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncmp.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mateoandre <mateoandre@student.42.fr>      +#+  +:+       +#+        */
+/*   By: mandre <mandre@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/14 10:08:38 by mateoandre        #+#    #+#             */
-/*   Updated: 2025/03/14 10:19:23 by mateoandre       ###   ########.fr       */
+/*   Updated: 2025/03/16 17:51:38 by mandre           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int ft_strncmp(const char *s1, const char *s2, int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int i;
+	size_t	i;
 
-    i = 0;
-    while ((s1[i] == s2[i] && i < n )|| !s1[i] || !s2[i])
-        i++;
-    return(s1[i] - s2[i]);
+	i = 0;
+	while ((s1[i] == s2[i] && i < n) || !s1[i] || !s2[i])
+		i++;
+	return (s1[i] - s2[i]);
 }
 
 // #include <stdio.h>
